+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and may be removed/ignored in a future release.
 + Support submission of workflows from json files (issue #926)
 + Support submission of workflows from stdin (issue #926)
 + Prometheus metrics and telemetry (issue #896) (@bbc88ks)
++ Detect and fail upon unknown fields during argo submit & lint (issue #892)
 * Remove installer/uninstaller (issue #928)
 * Update golang compiler to v1.10.3
 * Update k8s dependencies to v1.10 and client-go to v7.0